move routing class to helpers
